changes requested from review for pr
