+++ lapis/simulator.py
@@ -6,9 +6,10 @@ from usim.basics import Queue
 
 from lapis.drone import Drone
 from lapis.job import job_to_queue_scheduler
-from lapis.utility.monitor import Monitoring, collect_pool_cobald_statistics, \
-    collect_user_demand, collect_job_statistics, collect_drone_cobald_statistics, \
-    collect_resource_statistics, collect_pool_status, collect_configuration_information
+from lapis.monitor.general import user_demand, job_statistics, \
+    resource_statistics, pool_status, configuration_information
+from lapis.monitor import Monitoring
+from lapis.monitor.cobald import drone_statistics, pool_statistics
 
 
 class Simulator(object):
@@ -26,13 +27,13 @@ class Simulator(object):
 
     def enable_monitoring(self):
         self.monitoring = Monitoring(self)
-        self.monitoring.register_statistic(collect_user_demand)
-        self.monitoring.register_statistic(collect_job_statistics)
-        self.monitoring.register_statistic(collect_pool_cobald_statistics)
-        self.monitoring.register_statistic(collect_drone_cobald_statistics)
-        self.monitoring.register_statistic(collect_resource_statistics)
-        self.monitoring.register_statistic(collect_pool_status)
-        self.monitoring.register_statistic(collect_configuration_information)
+        self.monitoring.register_statistic(user_demand)
+        self.monitoring.register_statistic(job_statistics)
+        self.monitoring.register_statistic(pool_statistics)
+        self.monitoring.register_statistic(drone_statistics)
+        self.monitoring.register_statistic(resource_statistics)
+        self.monitoring.register_statistic(pool_status)
+        self.monitoring.register_statistic(configuration_information)
 
     def create_job_generator(self, job_input, job_reader):
         self._job_generators.append((job_input, job_reader))
